requirements.txt
Flask==3.0.0
pytest==8.2.0
prometheus-client==0.20.0
python-dotenv==1.0.1
